refactor: Split security.go into detection logic and printer functions
- security.go (948 → 569 lines): AuditClusterSecurity, auditPod, auditContainer,
  detectEnvironment, calculateIssueCounts, isExpectedPrivileged (core logic)
- security_printer.go (new, ~378 lines): PrintSecurityAudit, printClusterSummary,
  printDetailedFindings, printRecommendations, PrintSecurityAuditJSON (CLI output)
- Verified identical behavior: CLI text output, JSON output, dashboard /api/summary
- All 46 tests still pass
